internal/generator: move package to internal
Updates #37
