+++ features/fastreflection/new_field.go
@@ -1,7 +1,7 @@
 package fastreflection
 
 import (
-	"github.com/cosmos/cosmos-proto/generator"
+	"github.com/cosmos/cosmos-proto/internal/generator"
 	"google.golang.org/protobuf/compiler/protogen"
 	"google.golang.org/protobuf/reflect/protoreflect"
 )
