+++ features/fastreflection/type.go
@@ -2,7 +2,7 @@ package fastreflection
 
 import (
 	"fmt"
-	"github.com/cosmos/cosmos-proto/generator"
+	"github.com/cosmos/cosmos-proto/internal/generator"
 
 	"google.golang.org/protobuf/compiler/protogen"
 )
