+++ features/protoc/feature.go
@@ -1,7 +1,7 @@
 package protoc
 
 import (
-	"github.com/cosmos/cosmos-proto/generator"
+	"github.com/cosmos/cosmos-proto/internal/generator"
 	"google.golang.org/protobuf/compiler/protogen"
 )
