+++ internal/generator/helpers.go
